Read-only view for alerts if customer module

+++ hestia/src/main/java/hestia/prometheus/alert/DeleteAlertGroupAction.java
@@ -1,5 +1,6 @@
 package hestia.prometheus.alert;
 
+import hestia.HestiaWebapp;
 import hestia.base.HAction;
 
 public class DeleteAlertGroupAction extends HAction {
@@ -9,6 +10,9 @@ public class DeleteAlertGroupAction extends HAction {
         String env = ctx.pathParam("env");
         String id = ctx.pathParam("id");
 
+        if (HestiaWebapp.config.isCustomer()) {
+            throw new RuntimeException();
+        }
         AlertGroupDAO.delete(env, id);
 
         ctx.redirect("/alert/" + env);
