+++ hestia/src/main/java/hestia/prometheus/alert/rule/DeleteAlertRuleAction.java
@@ -1,5 +1,6 @@
 package hestia.prometheus.alert.rule;
 
+import hestia.HestiaWebapp;
 import hestia.base.HAction;
 import hestia.prometheus.alert.AlertGroupDAO;
 
@@ -11,6 +12,9 @@ public class DeleteAlertRuleAction extends HAction {
         String g = ctx.pathParam("g");
         String id = ctx.pathParam("id");
         
+        if (HestiaWebapp.config.isCustomer()) {
+            throw new RuntimeException();
+        }
         AlertGroupDAO.delete(env, g, id);
         
         ctx.redirect("/alert/" + env);
